requirements.txt
Django~=4.0.1
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.13.0